refactor and fix some issues after reviews and test the app
