+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/CategoryItem.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
-import com.sanaa.tudee_assistant.presentation.composable.CategoryThumbnail
+import com.sanaa.tudee_assistant.presentation.component.CategoryThumbnail
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.TudeeTheme
 import com.sanaa.tudee_assistant.presentation.model.CategoryUiState
