+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/button/PrimaryButton.kt
@@ -13,7 +13,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
@@ -30,12 +29,7 @@ fun PrimaryButton(
     onClick: () -> Unit = {},
     backgroundModifier: Modifier = when (enabled) {
         true -> Modifier.background(
-            brush = Brush.linearGradient(
-                listOf(
-                    Theme.color.primaryGradientStart,
-                    Theme.color.primaryGradientEnd
-                )
-            )
+            brush = Theme.color.primaryGradient,
         )
 
         false -> Modifier.background(color = Theme.color.disable)
