+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/button/utils/ButtonContent.kt
@@ -14,8 +14,7 @@ fun ButtonContent(
         true -> Theme.color.primary
         false -> Theme.color.disable
     },
-
-    ) {
+) {
 
     Text(
         text = label,
